Move xmodule_modifiers.py to openedx/core/lib/xblock_utils.py
