+++ lms/djangoapps/courseware/tests/test_module_render.py
@@ -1000,7 +1000,7 @@ class TestStaffDebugInfo(ModuleStoreTestCase):
             self.user,
             self.descriptor
         )
-        with patch('xmodule_modifiers.grade_histogram') as mock_grade_histogram:
+        with patch('openedx.core.lib.xblock_utils.grade_histogram') as mock_grade_histogram:
             mock_grade_histogram.return_value = []
             module = render.get_module(
                 self.user,
@@ -1022,7 +1022,7 @@ class TestStaffDebugInfo(ModuleStoreTestCase):
             max_grade=1,
             state="{}",
         )
-        with patch('xmodule_modifiers.grade_histogram') as mock_grade_histogram:
+        with patch('openedx.core.lib.xblock_utils.grade_histogram') as mock_grade_histogram:
             mock_grade_histogram.return_value = []
             module = render.get_module(
                 self.user,
