+++ lms/djangoapps/mobile_api/course_info/views.py
@@ -7,7 +7,7 @@ from rest_framework.response import Response
 
 from courseware.courses import get_course_about_section, get_course_info_section_module
 from static_replace import make_static_urls_absolute, replace_static_urls
-from xmodule_modifiers import get_course_update_items
+from openedx.core.lib.xblock_utils import get_course_update_items
 
 from ..utils import mobile_view, mobile_course_access
