+++ openedx/core/lib/xblock_utils.py
